refactor: create an animal subdirectory in the model directory
