+++ animal-chipping-front/src/app/model/animal/animal-create-request.ts
