+++ animal-chipping-front/src/app/model/animal/animal-request.ts
