+++ animal-chipping-front/src/app/model/animal/animal-update-request.ts
